Updating conda env and setup script.

+++ setup_project_env.sh
@@ -2,8 +2,16 @@
 
 set -e
 
+# Make sure we leave caller back from whence they called
+current_dir="$(pwd)"
+clean_up() {
+    cd "$current_dir"
+}
+trap clean_up EXIT
+
 # Get path to directory of this script
 project_dir="$( cd -P "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+cd "$project_dir"
 
 conda_env_path="${project_dir}/env/conda-env.yml"
 conda_env_name="SDSD-phyddle"
@@ -31,26 +39,44 @@ echo "Adding pydot_ng requirement via pip..."
 echo ""
 python -m pip install pydot_ng
 
-echo ""
-echo "Cloning SDSDsim..."
-echo ""
-git clone git@github.com:phyletica/SDSDsim.git
-(
-    cd SDSDsim
-    git checkout main
-    git pull
-    echo "Installing SDSDsim within cond env '$conda_env_name' using '$(which python)'"
-    python -m pip install -e .
-)
+if [ -e "SDSDsim" ]
+then
+    (
+        cd SDSDsim
+        echo "Installing existing SDSDsim within cond env '$conda_env_name' using '$(which python)'"
+        python -m pip install -e .
+    )
+else
+    echo ""
+    echo "Cloning SDSDsim..."
+    echo ""
+    git clone git@github.com:phyletica/SDSDsim.git
+    (
+        cd SDSDsim
+        git checkout main
+        git pull
+        echo "Installing SDSDsim within cond env '$conda_env_name' using '$(which python)'"
+        python -m pip install -e .
+    )
+fi
 
-echo ""
-echo "Cloning phyddle..."
-echo ""
-git clone git@github.com:mlandis/phyddle.git
-(
-    cd phyddle
-    git checkout main
-    git pull
-    echo "Installing phyddle within cond env '$conda_env_name' using '$(which python)'"
-    python -m pip install -e .
-)
+if [ -e "phyddle" ]
+then
+    (
+        cd phyddle
+        echo "Installing existing phyddle within cond env '$conda_env_name' using '$(which python)'"
+        python -m pip install -e .
+    )
+else
+    echo ""
+    echo "Cloning phyddle..."
+    echo ""
+    git clone git@github.com:mlandis/phyddle.git
+    (
+        cd phyddle
+        git checkout main
+        git pull
+        echo "Installing phyddle within cond env '$conda_env_name' using '$(which python)'"
+        python -m pip install -e .
+    )
+fi
